requirements.txt
asgiref==3.4.1
botocore==1.22.7
branca==0.4.2
certifi==2021.10.8
charset-normalizer==2.0.7
dj-database-url==0.5.0
Django==3.2.8
django-storages==1.12.3
folium==0.12.1
gcloud==0.18.3
googleapis-common-protos==1.53.0
httplib2==0.20.1
idna==3.3
Jinja2==3.0.2
jmespath==0.10.0
jws==0.1.3
MarkupSafe==2.0.1
numpy==1.21.3
oauth2client==4.1.3
Pillow==8.4.0
protobuf==3.19.1
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.11.0
pyparsing==2.4.7
Pyrebase4==4.5.0
python-dateutil==2.8.2
python-jwt==2.0.1
pytz==2021.3
requests==2.26.0
requests-toolbelt==0.9.1
rsa==4.7.2
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
